requirements.txt
python-docx==0.8.11
openai==1.51.0
python-dotenv==1.0.0
jieba==0.42.1
pypinyin==0.50.0
requests==2.31.0
pydantic==2.5.3
rich==13.7.0
click==8.1.7 